Refactor CLI into packs and harden Ctrl+C handling
