+++ axis/cli/packs/__init__.py
@@ -1,0 +1,1 @@
+"""Built-in CLI plugin packs."""
