+++ axis/cli/packs/events.py
@@ -1,0 +1,149 @@
+# ruff: noqa: D100,D103
+
+from __future__ import annotations
+
+import asyncio
+
+from aiohttp import ClientSession
+
+from axis.cli.packs.devices import (
+    DeviceEntry,
+    get_device_credentials,
+    run_on_selected_device,
+)
+from axis.device import AxisDevice
+from axis.errors import RequestError
+from axis.models.configuration import Configuration
+
+
+def register(registry: object, router: object) -> None:
+    """Register event-pack commands and menu nodes (explicit composition placeholder)."""
+
+
+async def fetch_event_instances(device_entry: DeviceEntry) -> list[dict[str, str]]:
+    async def _operation(device: AxisDevice) -> list[dict[str, str]]:
+        await device.vapix.event_instances.update()
+        return [
+            {
+                "topic": ev.topic,
+                "name": ev.name,
+                "stateful": str(ev.stateful),
+                "stateless": str(ev.stateless),
+                "available": str(ev.is_available),
+            }
+            for ev in device.vapix.event_instances.values()
+        ]
+
+    result = await run_on_selected_device(device_entry, _operation)
+    return result if result is not None else []
+
+
+def list_event_instances_flow(
+    serial: str, device_entry: DeviceEntry
+) -> list[dict[str, str]]:
+    events = asyncio.run(fetch_event_instances(device_entry))
+    if not events:
+        print(f"No event instances found for {serial}.")  # noqa: T201
+        return []
+
+    print(f"\nEvent instances for {serial}:")  # noqa: T201
+    for idx, ev in enumerate(events, 1):
+        flags = []
+        if ev["stateful"] == "True":
+            flags.append("stateful")
+        if ev["stateless"] == "True":
+            flags.append("stateless")
+        if ev["available"] == "True":
+            flags.append("available")
+        flag_str = ", ".join(flags) if flags else "none"
+        print(f"  {idx}. {ev['name']} [{ev['topic']}] ({flag_str})")  # noqa: T201
+    return events
+
+
+def _live_listen_flow(device_entry: DeviceEntry, topic_filter: str | None) -> None:
+    credentials = get_device_credentials(device_entry)
+    if credentials is None:
+        print("Stored device config is incomplete. Please re-add the device.")  # noqa: T201
+        return
+
+    if topic_filter:
+        print(f"\nListening for events on topic '{topic_filter}' (Ctrl+C to stop)...")  # noqa: T201
+    else:
+        print("\nListening for all events (Ctrl+C to stop)...")  # noqa: T201
+
+    async def _run_listener() -> None:
+        async with ClientSession() as session:
+            config = Configuration(
+                session=session,
+                host=credentials["host"],
+                username=credentials["username"],
+                password=credentials["password"],
+            )
+            device = AxisDevice(config)
+
+            def _on_event(event: object) -> None:
+                print(f"  [event] {event}")  # noqa: T201
+
+            unsubscribe = device.event.subscribe(
+                callback=_on_event,
+                id_filter=topic_filter,
+            )
+            stop_event = asyncio.Event()
+            try:
+                device.enable_events()
+                await stop_event.wait()
+            except asyncio.CancelledError:
+                pass
+            finally:
+                unsubscribe()
+
+    try:
+        asyncio.run(_run_listener())
+    except KeyboardInterrupt:
+        print("\nStopped listening.")  # noqa: T201
+    except RequestError as exc:
+        print(f"Device request failed: {exc}")  # noqa: T201
+
+
+def events_flow(serial: str, device_entry: DeviceEntry) -> None:
+    while True:
+        print(f"\nEvent options for {serial}:")  # noqa: T201
+        print("  1. List event instances")  # noqa: T201
+        print("  2. Listen to events (all topics)")  # noqa: T201
+        print("  3. Listen to events (select topic)")  # noqa: T201
+        print("  b. Back")  # noqa: T201
+        print("  e. Exit")  # noqa: T201
+        choice = input("Select option [1/2/3/b/e]: ").strip().lower()
+
+        if choice == "e":
+            print("Exiting.")  # noqa: T201
+            raise SystemExit(0)
+        if choice == "b":
+            return
+
+        if choice == "1":
+            list_event_instances_flow(serial, device_entry)
+            continue
+
+        if choice == "2":
+            _live_listen_flow(device_entry, topic_filter=None)
+            continue
+
+        if choice == "3":
+            events = list_event_instances_flow(serial, device_entry)
+            if not events:
+                continue
+            topic_choice = input("Enter event number to filter by: ").strip()
+            try:
+                topic_idx = int(topic_choice)
+            except ValueError:
+                print("Invalid selection.")  # noqa: T201
+                continue
+            if topic_idx < 1 or topic_idx > len(events):
+                print("Invalid selection.")  # noqa: T201
+                continue
+            selected_topic = events[topic_idx - 1]["topic"]
+            _live_listen_flow(device_entry, topic_filter=selected_topic)
+            continue
+
+        print("Invalid option. Please enter 1, 2, 3, b, or e.")  # noqa: T201
