+++ axis/cli/packs/navigation.py
@@ -1,0 +1,139 @@
+# ruff: noqa: D100,D103,TC003
+
+from __future__ import annotations
+
+import asyncio
+from pathlib import Path
+
+from axis.cli.packs.accounts import account_management_flow
+from axis.cli.packs.api import api_drill_down_flow, list_supported_apis_flow
+from axis.cli.packs.devices import (
+    DeviceEntry,
+    DeviceStore,
+    config_to_toml_dict,
+    find_serial_by_host,
+    get_config_path,
+    load_devices,
+    migrate_unknown_entry,
+    print_registered_devices,
+    prompt_for_device,
+    save_devices,
+    select_device,
+    validate_and_fetch_device,
+)
+from axis.cli.packs.events import events_flow
+
+
+def register(registry: object, router: object) -> None:
+    """Register navigation-pack commands and menu nodes (explicit composition placeholder)."""
+
+
+def selected_device_operations(serial: str, device_entry: DeviceEntry) -> None:
+    while True:
+        print(f"\nDevice operations for {serial}:")  # noqa: T201
+        print("  1. List supported APIs")  # noqa: T201
+        print("  2. API drill-down")  # noqa: T201
+        print("  3. Event instances & live listen")  # noqa: T201
+        print("  4. Account management")  # noqa: T201
+        print("  b. Back")  # noqa: T201
+        print("  e. Exit")  # noqa: T201
+        choice = input("Select option [1-4/b/e]: ").strip().lower()
+
+        if choice == "1":
+            list_supported_apis_flow(serial, device_entry)
+            continue
+
+        if choice == "2":
+            api_drill_down_flow(device_entry)
+            continue
+
+        if choice == "3":
+            events_flow(serial, device_entry)
+            continue
+
+        if choice == "4":
+            account_management_flow(serial, device_entry)
+            continue
+
+        if choice == "b":
+            return
+
+        if choice == "e":
+            print("Exiting.")  # noqa: T201
+            raise SystemExit(0)
+
+        print("Invalid option. Please enter 1, 2, 3, 4, b, or e.")  # noqa: T201
+
+
+def run_main_loop(config_path: Path | None = None) -> None:
+    resolved_config_path = config_path if config_path is not None else get_config_path()
+
+    while True:
+        devices: DeviceStore = load_devices(resolved_config_path)
+        print_registered_devices(devices)
+
+        print("\nOptions:")  # noqa: T201
+        print("  1. Add additional device")  # noqa: T201
+        print("  2. Device operations")  # noqa: T201
+        print("  e. Exit")  # noqa: T201
+        choice = input("Select an option [1/2/e]: ").strip().lower()
+
+        if choice == "e":
+            print("Exiting.")  # noqa: T201
+            raise SystemExit(0)
+
+        if choice == "2":
+            selected = select_device(devices)
+            if selected is None:
+                continue
+            selected_serial, selected_entry = selected
+            selected_device_operations(selected_serial, selected_entry)
+            continue
+
+        if choice != "1":
+            print("Invalid option. Please enter 1, 2, or e.")  # noqa: T201
+            continue
+
+        device_info = prompt_for_device()
+        existing_serial_for_host = find_serial_by_host(devices, device_info["host"])
+        if existing_serial_for_host is not None:
+            update_existing = (
+                input(
+                    f"A device with host {device_info['host']} already exists "
+                    f"(serial {existing_serial_for_host}). Update it? (y/n): "
+                )
+                .strip()
+                .lower()
+            )
+            if update_existing != "y":
+                print("Device registration aborted.")  # noqa: T201
+                continue
+
+        config, serial, model, extra = asyncio.run(
+            validate_and_fetch_device(device_info)
+        )
+        if config is None or serial is None or model is None or extra is None:
+            continue
+
+        migrate_unknown_entry(devices, serial, device_info["host"])
+
+        if serial in devices:
+            update = (
+                input(
+                    f"A device with serial {serial} already exists. "
+                    "Update its configuration? (y/n): "
+                )
+                .strip()
+                .lower()
+            )
+            if update != "y":
+                print("Device registration aborted.")  # noqa: T201
+                continue
+
+        devices[serial] = {
+            "config": config_to_toml_dict(config),
+            "model": model,
+            "extra": extra,
+        }
+        save_devices(resolved_config_path, devices)
+        print(f"Device '{serial}' registered/updated.")  # noqa: T201
